model: reorg model files
